+++ Tests/SentryProfilerTests/SentryTraceProfilerTests.swift
@@ -429,9 +429,7 @@ private extension SentryTraceProfilerTests {
     }
 
     func printTimestamps(entries: [[String: Any]]) -> [NSString] {
-        entries.reduce(into: [NSString](), { partialResult, entry in
-            partialResult.append(entry["elapsed_since_start_ns"] as! NSString)
-        })
+        entries.compactMap({ $0["elapsed_since_start_ns"] as? NSString })
     }
 
     func assertMetricEntries(measurements: [String: Any], key: String, expectedEntries: [[String: Any]], transaction: Transaction) throws {
@@ -466,10 +464,10 @@ private extension SentryTraceProfilerTests {
         XCTAssertEqual(values.count, numberOfReadings, "Wrong number of values under \(key)")
 
         if let expectedValue = expectedValue {
-            let actualValue = try XCTUnwrap(values[1]["value"] as? T)
+            let actualValue = try XCTUnwrap(values.element(at: 1)?["value"] as? T)
             XCTAssertEqual(actualValue, expectedValue, "Wrong value for \(key)")
 
-            let timestamp = try XCTUnwrap(values[0]["elapsed_since_start_ns"] as? NSString)
+            let timestamp = try XCTUnwrap(values.first?["elapsed_since_start_ns"] as? NSString)
             try assertTimestampOccursWithinTransaction(timestamp: timestamp, transaction: transaction)
 
             let actualUnits = try XCTUnwrap(metricContainer["unit"] as? String)
@@ -529,7 +527,7 @@ private extension SentryTraceProfilerTests {
         let debugMeta = try XCTUnwrap(profile["debug_meta"] as? [String: Any])
         let images = try XCTUnwrap(debugMeta["images"] as? [[String: Any]])
         XCTAssertFalse(images.isEmpty)
-        let firstImage = images[0]
+        let firstImage = try XCTUnwrap(images.first)
         XCTAssertFalse(try XCTUnwrap(firstImage["code_file"] as? String).isEmpty)
         XCTAssertFalse(try XCTUnwrap(firstImage["debug_id"] as? String).isEmpty)
         XCTAssertFalse(try XCTUnwrap(firstImage["image_addr"] as? String).isEmpty)
@@ -557,8 +555,8 @@ private extension SentryTraceProfilerTests {
 
         let frames = try XCTUnwrap(sampledProfile["frames"] as? [[String: Any]])
         XCTAssertFalse(frames.isEmpty)
-        XCTAssertFalse(try XCTUnwrap(frames[0]["instruction_addr"] as? String).isEmpty)
-        XCTAssertFalse(try XCTUnwrap(frames[0]["function"] as? String).isEmpty)
+        XCTAssertFalse(try XCTUnwrap(frames.first?["instruction_addr"] as? String).isEmpty)
+        XCTAssertFalse(try XCTUnwrap(frames.first?["function"] as? String).isEmpty)
 
         let stacks = try XCTUnwrap(sampledProfile["stacks"] as? [[Int]])
         var foundAtLeastOneNonEmptySample = false