+++ Tests/SentryTests/Networking/SentryTransportFactoryTests.swift
@@ -6,7 +6,7 @@ class SentryTransportFactoryTests: XCTestCase {
     
     private static let dsnAsString = TestConstants.dsnAsString(username: "SentryTransportFactoryTests")
 
-    func testIntegration_UrlSessionDelegate_PassedToRequestManager() {
+    func testIntegration_UrlSessionDelegate_PassedToRequestManager() throws {
         let urlSessionDelegateSpy = UrlSessionDelegateSpy()
         
         let expect = expectation(description: "UrlSession Delegate of Options called in RequestManager")
@@ -21,7 +21,7 @@ class SentryTransportFactoryTests: XCTestCase {
         let fileManager = try! SentryFileManager(options: options, dispatchQueueWrapper: TestSentryDispatchQueueWrapper())
         let transports = TransportInitializer.initTransports(options, sentryFileManager: fileManager, currentDateProvider: TestCurrentDateProvider())
         let httpTransport = transports.first
-        let requestManager = Dynamic(httpTransport).requestManager.asObject as! SentryQueueableRequestManager
+        let requestManager = try XCTUnwrap(Dynamic(httpTransport).requestManager.asObject as? SentryQueueableRequestManager)
         
         let imgUrl = URL(string: "https://github.com")!
         let request = URLRequest(url: imgUrl)
@@ -47,7 +47,7 @@ class SentryTransportFactoryTests: XCTestCase {
         let transports = TransportInitializer.initTransports(options, sentryFileManager: fileManager, currentDateProvider: TestCurrentDateProvider())
                 
         let httpTransport = transports.first
-        let requestManager = Dynamic(httpTransport).requestManager.asObject as! SentryQueueableRequestManager
+        let requestManager = try XCTUnwrap(Dynamic(httpTransport).requestManager.asObject as? SentryQueueableRequestManager)
         
         let imgUrl = URL(string: "https://github.com")!
         let request = URLRequest(url: imgUrl)