+++ Tests/SentryTests/Protocol/SentryClientReportTests.swift
@@ -10,7 +10,7 @@ class SentryClientReportTests: XCTestCase {
         clearTestState()
     }
 
-    func testSerialize() {
+    func testSerialize() throws {
         SentryDependencyContainer.sharedInstance().dateProvider = TestCurrentDateProvider()
         
         let event1 = SentryDiscardedEvent(reason: .sampleRate, category: .transaction, quantity: 2)
@@ -23,15 +23,15 @@ class SentryClientReportTests: XCTestCase {
         
         XCTAssertEqual(SentryDependencyContainer.sharedInstance().dateProvider.date().timeIntervalSince1970, actual["timestamp"] as? TimeInterval)
         
-        let discardedEvents = actual["discarded_events"] as! [[String: Any]]
+        let discardedEvents = try XCTUnwrap(actual["discarded_events"] as? [[String: Any]])
         
         func assertEvent(event: [String: Any], reason: String, category: String, quantity: UInt) {
             XCTAssertEqual(reason, event["reason"] as? String)
             XCTAssertEqual(category, event["category"] as? String)
             XCTAssertEqual(quantity, event["quantity"] as? UInt)
         }
-        assertEvent(event: discardedEvents[0], reason: "sample_rate", category: "transaction", quantity: event1.quantity)
-        assertEvent(event: discardedEvents[1], reason: "before_send", category: "transaction", quantity: event2.quantity)
-        assertEvent(event: discardedEvents[2], reason: "ratelimit_backoff", category: "error", quantity: event3.quantity)
+        assertEvent(event: try XCTUnwrap(discardedEvents.first), reason: "sample_rate", category: "transaction", quantity: event1.quantity)
+        assertEvent(event: try XCTUnwrap(discardedEvents.element(at: 1)), reason: "before_send", category: "transaction", quantity: event2.quantity)
+        assertEvent(event: try XCTUnwrap(discardedEvents.element(at: 2)), reason: "ratelimit_backoff", category: "error", quantity: event3.quantity)
     }
 }