+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Include the screen names in the session replay (#4126)
 
+### Fixes
+
+- Sentry Replay Serialized Breadcrumbs include level name ([#4141](https://github.com/getsentry/sentry-cocoa/pull/4141))
+
 ## 8.30.0
 
 ### Features